requirements.txt
beautifulsoup4==4.10.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
Flask==3.0.3
fpdf==1.7.2
gunicorn==23.0.0
idna==3.8
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
newspaper3k==0.2.8
nltk==3.9.1
packaging==24.1
pillow==10.4.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
urllib3==2.2.2
Werkzeug==3.0.4
lxml==5.3.0
lxml_html_clean==0.2.1 
